refactor(commands): extract /copy command into copy.rs
